trend-scout/requirements.txt
# Composio + LangChain
composio-langchain>=0.5.0
composio-core>=0.5.0

# LLM
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# Web search fallback (for X/Twitter without API)
tavily-python>=0.5.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Database
sqlalchemy>=2.0.0

# UI
streamlit>=1.38.0

# Utils
python-dotenv>=1.0.0
httpx>=0.27.0
rich>=13.0.0
